+++ src/Seo.php
@@ -16,7 +16,7 @@ use craft\web\twig\variables\CraftVariable;
 use craft\web\UrlManager;
 use craft\web\View;
 use ether\seo\fields\SeoField;
-use ether\seo\listeners\GetCraftQLSchema;
+use ether\seo\integrations\craftql\GetCraftQLSchema;
 use ether\seo\models\Settings;
 use ether\seo\services\RedirectsService;
 use ether\seo\services\SeoService;
@@ -150,8 +150,7 @@ class Seo extends Plugin
 		}
 
 		// CraftQL Support
-		/** @noinspection PhpUndefinedNamespaceInspection */
-		/** @noinspection PhpUnnecessaryFullyQualifiedNameInspection */
+		/** @noinspection PhpFullyQualifiedNameUsageInspection */
 		if (class_exists(\markhuot\CraftQL\CraftQL::class)) {
 			Event::on(
 				SeoField::class,